+++ build.gradle
@@ -32,7 +32,7 @@ ext {
     lintAbortOnError = false
 
     // QuickBlox SDK version
-    qbSdkVersion = '3.3.1'
+    qbSdkVersion = '3.3.3'
 
     versionName = '3.3.0'
