+++ sample-core/src/main/java/com/quickblox/sample/core/CoreApp.java
@@ -4,6 +4,9 @@ import android.app.Application;
 import android.text.TextUtils;
 import android.util.Log;
 
+import com.quickblox.auth.session.QBSession;
+import com.quickblox.auth.session.QBSessionManager;
+import com.quickblox.auth.session.QBSessionParameters;
 import com.quickblox.auth.session.QBSettings;
 import com.quickblox.core.ServiceZone;
 import com.quickblox.sample.core.models.QbConfigs;
@@ -20,6 +23,7 @@ public class CoreApp extends Application {
     public void onCreate() {
         super.onCreate();
         instance = this;
+        initQBSessionManager();
         initQbConfigs();
         initCredentials();
     }
@@ -51,5 +55,39 @@ public class CoreApp extends Application {
 
     protected String getQbConfigFileName(){
         return QB_CONFIG_DEFAULT_FILE_NAME;
+    }
+
+    private void initQBSessionManager() {
+        QBSessionManager.getInstance().addListener(new QBSessionManager.QBSessionListener() {
+            @Override
+            public void onSessionCreated(QBSession qbSession) {
+                Log.d(TAG, "Session Created");
+            }
+
+            @Override
+            public void onSessionUpdated(QBSessionParameters qbSessionParameters) {
+                Log.d(TAG, "Session Updated");
+            }
+
+            @Override
+            public void onSessionDeleted() {
+                Log.d(TAG, "Session Deleted");
+            }
+
+            @Override
+            public void onSessionRestored(QBSession qbSession) {
+                Log.d(TAG, "Session Restored");
+            }
+
+            @Override
+            public void onSessionExpired() {
+                Log.d(TAG, "Session Expired");
+            }
+
+            @Override
+            public void onProviderSessionExpired(String provider) {
+                Log.d(TAG, "Session Expired for provider:" + provider);
+            }
+        });
     }
 }
