--- sample-location/README.md
@@ -1,18 +1,0 @@
-## QuickBlox Android Location Sample
-
-This is a code sample for [QuickBlox](http://quickblox.com/) platform. It is a great way for developers using QuickBlox platform to learn how to integrate location based features into yours apps.
-
-This is an Android build, it demonstrates how to work with [Location](http://quickblox.com/developers/Location) QuickBlox module.
-It allows to see other users on the Map and shares own position.
-
-It shows how to:
-<ul>
-<li> Retrieve all users’ points and show them on the Map</li>
-<li> Share own location for other users </li>
-</ul>
-
-Go to project page on QuickBlox Developers Section -- <http://quickblox.com/developers/SimpleSample-location_users-android>
-
-Or clone project here and start to work with the code.
-
-<img src="http://files.quickblox.com/Sample_Location_2.png" width=200 />&nbsp;<img src="http://files.quickblox.com/Sample_Location_1.png" width=200 />
