--- sample-location/build.gradle
@@ -1,54 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion rootProject.compileSdkVersion
-    buildToolsVersion rootProject.buildToolsVersion
-
-    defaultConfig {
-        applicationId "com.quickblox.sample.location"
-        minSdkVersion rootProject.minSdkVersion
-        targetSdkVersion rootProject.targetSdkVersion
-        versionCode 1
-        versionName "1.0"
-    }
-
-    productFlavors {
-        speedDev {
-            minSdkVersion 21
-        }
-        dev {
-            minSdkVersion rootProject.minSdkVersion
-        }
-    }
-
-    buildTypes {
-        debug {
-            signingConfig signingConfigs.debug
-        }
-    }
-
-    signingConfigs {
-        debug {
-            storeFile file("../cert/debug.keystore")
-            storePassword "android"
-            keyAlias "androiddebugkey"
-            keyPassword "android"
-        }
-    }
-
-    lintOptions {
-        abortOnError false
-    }
-}
-
-dependencies {
-    compile "com.quickblox:quickblox-android-sdk-location:$rootProject.qbSdkVersion"
-    compile (project(":sample-core")) {
-        transitive = true;
-    }
-
-    compile "com.google.android.gms:play-services-location:${rootProject.playServicesVersion}"
-    compile "com.google.android.gms:play-services-maps:${rootProject.playServicesVersion}"
-}
-
-apply from: "../artifacts.gradle"
