--- sample-location/src/main/assets/qb_config.json
@@ -1,10 +1,0 @@
-{
-  "app_id": "99",
-  "auth_key": "63ebrp5VZt7qTOv",
-  "auth_secret": "YavMAxm5T59-BRw",
-  "account_key": "rz2sXxBt5xgSxGjALDW6",
-  "api_domain": "https://api.quickblox.com",
-  "chat_domain": "chat.quickblox.com",
-  "gcm_sender_id": "761750217637"
-}
-
