--- sample-location/src/main/assets/sample_config.json
@@ -1,4 +1,0 @@
-{
-  "user_login": "bobbobbob",
-  "user_password": "bobbobbob"
-}
