--- sample-location/src/main/java/com/quickblox/sample/location/App.java
@@ -1,11 +1,0 @@
-package com.quickblox.sample.location;
-
-import com.quickblox.sample.core.CoreApp;
-
-public class App extends CoreApp {
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-    }
-}
