--- sample-location/src/main/java/com/quickblox/sample/location/utils/Consts.java
@@ -1,12 +1,0 @@
-package com.quickblox.sample.location.utils;
-
-public class Consts {
-
-    public static final String USER_LOGIN_FIELD_NAME = "user_login";
-    public static final String USER_PASSWORD_FIELD_NAME = "user_password";
-    public static final String SAMPLE_CONFIG_FILE_NAME = "sample_config.json";
-
-    public static int LOCATION_MIN_TIME = 20000;
-    public static int LOCATION_PER_PAGE = 100;
-    public static int PLAY_SERVICE_REQUEST_CODE = 10;
-}
