--- sample-location/src/main/res/values/dimens.xml
@@ -1,15 +1,0 @@
-<resources>
-    <dimen name="margin_small">5dp</dimen>
-    <dimen name="margin_middle">10dp</dimen>
-    <dimen name="margin_large">20dp</dimen>
-
-    <dimen name="padding_small">5dp</dimen>
-    <dimen name="padding_middle">10dp</dimen>
-    <dimen name="padding_large">15dp</dimen>
-
-    <dimen name="logo_maxwidth">320dp</dimen>
-
-    <dimen name="text_size_small">15sp</dimen>
-    <dimen name="text_size_middle">18sp</dimen>
-    <dimen name="text_size_large">20sp</dimen>
-</resources>
