+++ sample-users/src/main/java/com/quickblox/sample/user/App.java
@@ -1,10 +1,5 @@
 package com.quickblox.sample.user;
 
-import android.util.Log;
-
-import com.quickblox.auth.session.QBSession;
-import com.quickblox.auth.session.QBSessionManager;
-import com.quickblox.auth.session.QBSessionParameters;
 import com.quickblox.sample.core.CoreApp;
 
 public class App extends CoreApp {
@@ -13,35 +8,6 @@ public class App extends CoreApp {
     @Override
     public void onCreate() {
         super.onCreate();
-        initQBSessionManager();
     }
 
-    private void initQBSessionManager() {
-        QBSessionManager.getInstance().addListener(new QBSessionManager.QBSessionListener() {
-            @Override
-            public void onSessionCreated(QBSession qbSession) {
-                Log.d(TAG, "Session Created");
-            }
-
-            @Override
-            public void onSessionUpdated(QBSessionParameters qbSessionParameters) {
-                Log.d(TAG, "Session Updated");
-            }
-
-            @Override
-            public void onSessionDeleted() {
-                Log.d(TAG, "Session Deleted");
-            }
-
-            @Override
-            public void onSessionRestored(QBSession qbSession) {
-                Log.d(TAG, "Session Restored");
-            }
-
-            @Override
-            public void onSessionExpired() {
-                Log.d(TAG, "Session Expired");
-            }
-        });
-    }
 }
