+++ sample-videochat-webrtc/src/main/java/com/quickblox/sample/groupchatwebrtc/fragments/BaseConversationFragment.java
@@ -23,6 +23,7 @@ import com.quickblox.sample.groupchatwebrtc.utils.Consts;
 import com.quickblox.sample.groupchatwebrtc.utils.UsersUtils;
 import com.quickblox.sample.groupchatwebrtc.utils.WebRtcSessionManager;
 import com.quickblox.users.model.QBUser;
+import com.quickblox.videochat.webrtc.BaseSession;
 import com.quickblox.videochat.webrtc.QBRTCSession;
 import com.quickblox.videochat.webrtc.QBRTCTypes;
 
@@ -140,7 +141,7 @@ public abstract class BaseConversationFragment extends BaseToolBarFragment imple
 
         }
 
-        if (currentSession.getState() != QBRTCSession.QBRTCSessionState.QB_RTC_SESSION_ACTIVE) {
+        if (currentSession.getState() != QBRTCSession.QBRTCSessionState.QB_RTC_SESSION_CONNECTED) {
             if (isIncomingCall) {
                 currentSession.acceptCall(null);
             } else {
